EWOK-78 EWOK-79 Initializing storyline manager in map activity. Moved/remove existing code related to polylines
